revert HttpClientProperties move to http folder
